Pin mycli version 1.10.0
Later versions of mycli run slowly in Docker. It seems like the query finishes on the
server in the usual amount of time, but the display of the results takes a few seconds.
However, version 1.10.0 displays results quickly.

mycli version 1.10.0 also requires pymysql 0.9.2, which is pinned here. pymysql 0.9.3
removes a function that mycli 1.10.0 relies on.